+++ src/components/notFound/NotFound.vue
@@ -1,11 +1,11 @@
 <template>
-  <section class="container container_m">
+  <b-container size="m">
     <h1>{{ $t('not_found') }}</h1>
     <p>{{ $t('not_found_desc') }}</p>
     <router-link :to="{ name: 'index' }">
       {{ $t('go_to_homepage') }}
     </router-link>
-  </section>
+  </b-container>
 </template>
 
 <script>