+++ src/components/search/AuthorList.vue
@@ -1,45 +1,28 @@
 <template>
-  <section class="container container_m">
-    <div class="card">
-      <div class="card_item" v-for="author in authors" :key="author.id">
-        <ul class="card_options" v-if="isAdmin">
-          <li class="card_option">
-            <button
-              class="card_option_btn"
-              :title="$t('remove')"
-              @click="remove(author.id)"
-            >
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                width="30"
-                height="30"
-                viewBox="0 0 50 50"
-                class="icon"
-              >
-                <path
-                  d="M11.2 2v8H2v6h46v-6h-7.36V2zm3.68 4h22.08v4H14.88zM7.52 18v30h34.96V18z"
-                />
-              </svg>
-            </button>
-          </li>
-        </ul>
-        <div class="card_text">
-          <router-link :to="{ name: 'edit-author', params: { id: author.id } }">
-            {{ formatAuthor(author) }}
-          </router-link>
-        </div>
-      </div>
-    </div>
-  </section>
+  <b-container size="m">
+    <b-list v-for="author in authors" :key="author.id">
+      <template #title>
+        <router-link :to="{ name: 'edit-author', params: { id: author.id } }">
+          {{ formatAuthor(author) }}
+        </router-link>
+      </template>
+      <template #options>
+        <b-dropdown-item>
+          <b-button type="text" class="dropdown_btn" @click="remove(author.id)">
+            {{ $t('remove') }}
+          </b-button>
+        </b-dropdown-item>
+      </template>
+    </b-list>
+  </b-container>
 </template>
 
 <script>
-import { mapState, mapGetters, mapActions } from 'vuex'
+import { mapState, mapActions } from 'vuex'
 
 export default {
   name: 'author',
   computed: {
-    ...mapGetters('user', ['isAdmin']),
     ...mapState('search', ['authors'])
   },
   methods: {