+++ src/components/staff/Edit.vue
@@ -1,27 +1,25 @@
 <template>
-  <div class="card_text" @dblclick="toggleEditStaff()" v-if="!editStaff">
+  <span @dblclick="toggleEditStaff()" v-if="!editStaff">
     {{ staff.name }}
-  </div>
-  <div class="card_text" v-else>
-    <form class="form form_inline" @submit.prevent="update()">
-      <div class="form_group">
-        <div class="form_item">
-          <label for="staff" class="form_label visuallyHidden">
-            {{ $t('staff') }}
-          </label>
-        </div>
-        <div class="form_item">
-          <input
-            type="text"
-            id="staff"
-            class="form_input"
-            v-focus
-            v-model="name"
-          />
-        </div>
+  </span>
+  <b-form @submit.prevent="update()" v-else>
+    <div class="form_group">
+      <div class="form_item">
+        <label for="staff" class="form_label visuallyHidden">
+          {{ $t('staff') }}
+        </label>
       </div>
-    </form>
-  </div>
+      <div class="form_item">
+        <input
+          type="text"
+          id="staff"
+          class="form_input"
+          v-model="name"
+          v-focus
+        />
+      </div>
+    </div>
+  </b-form>
 </template>
 
 <script>
@@ -55,9 +53,3 @@ export default {
   }
 }
 </script>
-
-<style>
-.form_inline {
-  margin: 0;
-}
-</style>