+++ src/components/staff/List.vue
@@ -1,32 +1,18 @@
 <template>
-  <section class="container container_m">
-    <div class="card">
-      <div class="card_item" v-for="member in staff" :key="member.id">
-        <ul class="card_options" v-if="isAdmin">
-          <li class="card_option">
-            <button
-              class="card_option_btn"
-              :title="$t('remove')"
-              @click="remove(member.id)"
-            >
-              <svg
-                xmlns="http://www.w3.org/2000/svg"
-                width="30"
-                height="30"
-                viewBox="0 0 50 50"
-                class="icon"
-              >
-                <path
-                  d="M11.2 2v8H2v6h46v-6h-7.36V2zm3.68 4h22.08v4H14.88zM7.52 18v30h34.96V18z"
-                />
-              </svg>
-            </button>
-          </li>
-        </ul>
+  <b-container size="m">
+    <b-list v-for="member in staff" :key="member.id">
+      <template #options>
+        <b-dropdown-item v-if="isAdmin">
+          <b-button type="text" class="dropdown_btn" @click="remove(member.id)">
+            {{ $t('remove') }}
+          </b-button>
+        </b-dropdown-item>
+      </template>
+      <template #title>
         <edit-staff :staff="member" />
-      </div>
-    </div>
-  </section>
+      </template>
+    </b-list>
+  </b-container>
 </template>
 
 <script>